requirements.txt
defusedxml==0.4.1
factory-boy==2.7.0
fake-factory==0.5.10
mock==2.0.0
pytest==2.9.2
pytest-django==2.9.1
pytest-pythonpath==0.7
requests==2.10.0
